% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{myboot}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05)
}
\arguments{
\item{iter}{interations}

\item{x}{sample}

\item{fun}{function}

\item{alpha}{alpha}
}
\value{
none
}
\description{
myboot
}
