% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iteration}

\item{x}{sample}

\item{fun}{function}

\item{alpha}{alpha}

\item{cx}{cex}

\item{...}{additional code for the histogram}
}
\value{
histogram
}
\description{
myboot2
}
\examples{
\dontrun{iter=10000,x,fun="mean",alpha=0.05,cx=1.5}
}
