% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n = 100, iter = 1000, a = 0, b = 5)
}
\arguments{
\item{n}{number of samples}

\item{iter}{number of iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
a histogram of the central tendency theorem
}
\description{
myclt
}
\examples{
\dontrun{mycltb(n= 10, iter = 10000, a = 0, b = 5}
}
