% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{my f}
\usage{
myf(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iteration}

\item{n}{sample size}

\item{p}{probabily}
}
\value{
barplot of the proportions
}
\description{
my f
}
\examples{
\dontrun{myf(iter=1000, n= 10, p=0.5)}
}
