% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function}

\item{x}{data}

\item{param}{function to be utilized}
}
\value{
plot and a list of i, param, y, and slope
}
\description{
mymaxlik
}
