% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymld.R
\name{mymld}
\alias{mymld}
\title{mymld}
\usage{
mymld(lfun, x, param, delta = 1e-04)
}
\arguments{
\item{lfun}{function}

\item{x}{data}

\item{param}{parameter values}

\item{delta}{accurancy}
}
\description{
mymld
}
