% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{median}

\item{sigma}{standard deviation}

\item{a}{area}
}
\value{
a list of the median, standard deviation and area
}
\description{
myncurve
}
\examples{
\dontrun{myncurve(mu = 3, sigma = 5, a = 10)}
}
