% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{myread}
\usage{
myread(csv, dird)
}
\arguments{
\item{csv}{a file from t}

\item{dird}{the file location}
}
\value{
a vector with the data from the csv file
}
\description{
myread
}
