% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{number of seats on plane}

\item{gamma}{probability of overbooking}

\item{p}{probability of show}
}
\value{
two plots of discrete and continuous distributions and a list of all values
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N= 200, gamma = 0.02, p = 0.95}
}
