% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats}

\item{gamma}{Probability a plane will be truly overbooked}

\item{p}{Probability of a "show"}
}
\value{
A list containing nd, nc, N, p and gamma
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N=400,gamma = 0.02, p = 0.95)}
}
